auth module refactoring
