+++ src/store/redux/store.ts
@@ -4,8 +4,7 @@ import createSagaMiddleware from "redux-saga";
 import { fork } from "redux-saga/effects";
 import { gameFieldSlice } from "@/components/GameField";
 import { fieldControlSlice, fieldControlsSaga } from "@/components/Fields";
-import { authSlice } from "@/auth";
-import { authSaga } from "@/auth/AuthRdx";
+import { authSlice, authSaga } from "@/auth";
 
 const sagaMiddleware = createSagaMiddleware();
