+++ src/utils/test-utils.tsx
@@ -2,12 +2,8 @@ import React, { ReactElement } from "react";
 import { render as rtlRender } from "@testing-library/react";
 import { configureStore } from "@reduxjs/toolkit";
 import { Provider } from "react-redux";
-// Import your own reducer
-import { authSaga, authSlice } from "@/auth/AuthRdx";
-import {
-  fieldControlSlice,
-  fieldControlsSaga,
-} from "@/components/Fields/FieldControlRdx";
+import { authSlice, authSaga } from "@/auth";
+import { fieldControlSlice, fieldControlsSaga } from "@/components/Fields";
 import { gameFieldSlice } from "@/components/GameField/GameFieldRdx";
 import { initialStateAll } from "@/store/redux/store";
 import createSagaMiddleware from "redux-saga";
